--- src/main/java/com/isthisteamisthis/lalalia/scoresongdata/query/application/controller/test.java
@@ -1,4 +1,0 @@
-package com.isthisteamisthis.lalalia.scoresongdata.query.application.controller;
-
-public class test {
-}
