--- src/main/java/com/isthisteamisthis/lalalia/scoresongdata/query/application/dto/test.java
@@ -1,3 +1,0 @@
-package com.isthisteamisthis.lalalia.scoresongdata.query.application.dto;
-public class test {
-}
