--- src/main/java/com/isthisteamisthis/lalalia/scoresongdata/query/domain/aggregate/test.java
@@ -1,3 +1,0 @@
-package com.isthisteamisthis.lalalia.scoresongdata.query.domain.aggregate;
-public class test {
-}
