--- src/main/java/com/isthisteamisthis/lalalia/scoresongdata/query/domain/repository/test.java
@@ -1,3 +1,0 @@
-package com.isthisteamisthis.lalalia.scoresongdata.query.domain.repository;
-public class test {
-}
