--- src/main/java/com/isthisteamisthis/lalalia/scoresongdata/query/infrastructure/repository/test.java
@@ -1,4 +1,0 @@
-package com.isthisteamisthis.lalalia.scoresongdata.query.infrastructure.repository;
-
-public class test {
-}
