--- src/main/java/com/isthisteamisthis/lalalia/scoresongdata/query/infrastructure/service/test.java
@@ -1,4 +1,0 @@
-package com.isthisteamisthis.lalalia.scoresongdata.query.infrastructure.service;
-
-public class test {
-}
