+++ src/main/java/com/isthisteamisthis/lalalia/songdata/query/infrastructure/repository/test.java
@@ -1,0 +1,4 @@
+package com.isthisteamisthis.lalalia.songdata.query.infrastructure.repository;
+
+public class test {
+}
