+++ src/main/java/com/isthisteamisthis/lalalia/songdata/query/infrastructure/service/test.java
@@ -1,0 +1,4 @@
+package com.isthisteamisthis.lalalia.songdata.query.infrastructure.service;
+
+public class test {
+}
