--- src/test/java/com/isthisteamisthis/lalalia/rangesongdata/command/application/controller/test.java
@@ -1,4 +1,0 @@
-package com.isthisteamisthis.lalalia.rangesongdata.command.application.controller;
-
-public class test {
-}
