--- src/test/java/com/isthisteamisthis/lalalia/rangesongdata/command/application/service/test.java
@@ -1,3 +1,0 @@
-package com.isthisteamisthis.lalalia.rangesongdata.command.application.service;
-public class test {
-}
