--- src/test/java/com/isthisteamisthis/lalalia/rangesongdata/query/application/controller/test.java
@@ -1,4 +1,0 @@
-package com.isthisteamisthis.lalalia.rangesongdata.query.application.controller;
-
-public class test {
-}
