--- src/test/java/com/isthisteamisthis/lalalia/rangesongdata/query/application/dto/test.java
@@ -1,3 +1,0 @@
-package com.isthisteamisthis.lalalia.rangesongdata.query.application.dto;
-public class test {
-}
