--- src/test/java/com/isthisteamisthis/lalalia/rangesongdata/query/application/service/test.java
@@ -1,3 +1,0 @@
-package com.isthisteamisthis.lalalia.rangesongdata.query.application.service;
-public class test {
-}
