--- src/test/java/com/isthisteamisthis/lalalia/rangesongdata/query/domain/aggregate/test.java
@@ -1,3 +1,0 @@
-package com.isthisteamisthis.lalalia.rangesongdata.query.domain.aggregate;
-public class test {
-}
