--- src/test/java/com/isthisteamisthis/lalalia/rangesongdata/query/domain/repository/test.java
@@ -1,3 +1,0 @@
-package com.isthisteamisthis.lalalia.rangesongdata.query.domain.repository;
-public class test {
-}
