+++ src/test/java/com/isthisteamisthis/lalalia/songdata/command/application/controller/test.java
@@ -1,0 +1,4 @@
+package com.isthisteamisthis.lalalia.songdata.command.application.controller;
+
+public class test {
+}
