+++ src/test/java/com/isthisteamisthis/lalalia/songdata/command/application/service/test.java
@@ -1,0 +1,3 @@
+package com.isthisteamisthis.lalalia.songdata.command.application.service;
+public class test {
+}
