+++ src/test/java/com/isthisteamisthis/lalalia/songdata/query/application/dto/test.java
@@ -1,0 +1,3 @@
+package com.isthisteamisthis.lalalia.songdata.query.application.dto;
+public class test {
+}
