+++ src/test/java/com/isthisteamisthis/lalalia/songdata/query/application/service/test.java
@@ -1,0 +1,3 @@
+package com.isthisteamisthis.lalalia.songdata.query.application.service;
+public class test {
+}
