+++ src/test/java/com/isthisteamisthis/lalalia/songdata/query/domain/aggregate/test.java
@@ -1,0 +1,3 @@
+package com.isthisteamisthis.lalalia.songdata.query.domain.aggregate;
+public class test {
+}
