+++ src/test/java/com/isthisteamisthis/lalalia/songdata/query/domain/repository/test.java
@@ -1,0 +1,3 @@
+package com.isthisteamisthis.lalalia.songdata.query.domain.repository;
+public class test {
+}
